analog dice
These are python programs of analog dice.
The file die.py defines a class representing a single die(6 sides), and a method that return a random value between 1 and number of sides.
The file die_visual.py creates a D6, making some rolls and visualizing results.
The dice_visual.py creates two D6, making some rolls and visualizing results.
The different_dice.py creates a D6 and a D10, making some rolls and visualizing results.
And files .svg are the results of plots.